+++ internal/fleet/integration_policy/schema.go
@@ -99,6 +99,12 @@ func getSchemaV1() schema.Schema {
 				Optional:    true,
 				Sensitive:   true,
 			},
+			"space_ids": schema.SetAttribute{
+				Description: "The Kibana space IDs where this integration policy is available. When set, must match the space_ids of the referenced agent policy. If not set, will be inherited from the agent policy. Note: The order of space IDs does not matter as this is a set.",
+				ElementType: types.StringType,
+				Optional:    true,
+				Computed:    true,
+			},
 		},
 		Blocks: map[string]schema.Block{
 			"input": schema.ListNestedBlock{