+++ internal/fleet/output/acc_test.go
@@ -541,7 +541,7 @@ func checkResourceOutputDestroy(s *terraform.State) error {
 		if err != nil {
 			return err
 		}
-		output, diags := fleet.GetOutput(context.Background(), fleetClient, rs.Primary.ID)
+		output, diags := fleet.GetOutput(context.Background(), fleetClient, rs.Primary.ID, "")
 		if diags.HasError() {
 			return diagutil.FwDiagsAsError(diags)
 		}